+++ src/bsk_rl/data/resource_data.py
@@ -5,9 +5,10 @@ configured by setting the ``resource_fn`` argument to the desired function of th
 """
 
 import logging
-from typing import TYPE_CHECKING, Callable
+from typing import Callable, Union
 
 from bsk_rl.data.base import Data, DataStore, GlobalReward
+from bsk_rl.obs import ResourceRewardWeight
 
 logger = logging.getLogger(__name__)
 
@@ -57,7 +58,9 @@ class ResourceReward(GlobalReward):
     data_store_type = ResourceDataStore
 
     def __init__(
-        self, reward_weight: float = 1.0, resource_fn: Callable = lambda sat: 0.0
+        self,
+        reward_weight: Union[float, Callable] = 1.0,
+        resource_fn: Callable = lambda sat: 0.0,
     ) -> None:
         """Rewards for an arbitrary resource.
 
@@ -76,16 +79,32 @@ class ResourceReward(GlobalReward):
             resource_fn = lambda sat: sat.simulator.sim_time
             reward_weight = -1e-3  # is negative, because time increases over time
 
-
         Args:
             reward_weight: [reward/resource] Scaling factor to apply to changes in resource
-                level to yield reward.
+                level to yield reward. Can be a float or a function that randomizes the reward
+                weight per-episode.
             resource_fn: Function to call to get the resource level for each satellite.
         """
         super().__init__()
-        self.reward_weight = reward_weight
+        self._reward_weight = reward_weight
         self.resource_fn = resource_fn
         self.data_store_kwargs = dict(resource_fn=resource_fn)
+
+    def reset_pre_sim_init(self) -> None:
+        """Reset the reward weight before simulation initialization."""
+        if callable(self._reward_weight):
+            self.reward_weight = self._reward_weight()
+        else:
+            self.reward_weight = self._reward_weight
+        return super().reset_pre_sim_init()
+
+    def reset_post_sim_init(self) -> None:
+        """Add the reward weight to each satellite's observation spec."""
+        for satellite in self.scenario.satellites:
+            for obs in satellite.observation_builder.observation_spec:
+                if isinstance(obs, ResourceRewardWeight):
+                    obs.weight_vector.append(self.reward_weight)
+        return super().reset_post_sim_init()
 
     def calculate_reward(
         self, new_data_dict: dict[str, ResourceData]